+++ sdk/typescript/src/apis/index.ts
@@ -1,0 +1,7 @@
+/* tslint:disable */
+/* eslint-disable */
+export * from './AccountsApi';
+export * from './ConsentsApi';
+export * from './PaymentsApi';
+export * from './ProvidersApi';
+export * from './WebhooksApi';