+++ sdk/typescript/src/models/index.ts
@@ -1,0 +1,16 @@
+/* tslint:disable */
+/* eslint-disable */
+export * from './Account';
+export * from './AccountIdentifiersInner';
+export * from './Balance';
+export * from './Consent';
+export * from './CreateConsentRequest';
+export * from './CreatePaymentRequest';
+export * from './CreatePaymentRequestCreditor';
+export * from './CreateWebhookSubscriptionRequest';
+export * from './ModelError';
+export * from './Payment';
+export * from './Provider';
+export * from './Transaction';
+export * from './TransactionPage';
+export * from './WebhookSubscription';